Rename to test folder to avoid collisions.
