+++ tests/in_memory_persistence_layer_t/test_get_options.py
@@ -1,6 +1,6 @@
 
 from models.option import Option
-from tests.in_memory_persistence_layer.in_memory_test_base import \
+from tests.in_memory_persistence_layer_t.in_memory_test_base import \
     InMemoryTestBase
