+++ tests/in_memory_persistence_layer_t/test_get_paginated_tasks.py
@@ -2,7 +2,7 @@ import logging
 
 from models.tag import Tag
 from models.task import Task
-from tests.in_memory_persistence_layer.in_memory_test_base import \
+from tests.in_memory_persistence_layer_t.in_memory_test_base import \
     InMemoryTestBase
