+++ tests/in_memory_persistence_layer_t/test_get_tag.py
@@ -1,6 +1,6 @@
 
 from models.tag import Tag
-from tests.in_memory_persistence_layer.in_memory_test_base import \
+from tests.in_memory_persistence_layer_t.in_memory_test_base import \
     InMemoryTestBase
