+++ tests/in_memory_persistence_layer_t/test_get_task.py
@@ -1,5 +1,5 @@
 from models.task import Task
-from tests.in_memory_persistence_layer.in_memory_test_base import \
+from tests.in_memory_persistence_layer_t.in_memory_test_base import \
     InMemoryTestBase
